+++ src/core/lombok/javac/handlers/HandleBuilder.java
@@ -126,19 +126,19 @@ public class HandleBuilder extends JavacAnnotationHandler<Builder> {
 		if (builderMethodName == null) builderMethodName = "builder";
 		if (buildMethodName == null) buildMethodName = "build";
 		if (builderClassName == null) builderClassName = "";
-		
-		boolean generateBuilderMethod;
-		if (builderMethodName.isEmpty()) {
-			generateBuilderMethod = false;
-		} else if (!checkName("builderMethodName", builderMethodName, annotationNode)) {
+
+		boolean generateBuilderMethod = !builderMethodName.isEmpty();
+		if (generateBuilderMethod && !checkName("builderMethodName", builderMethodName, annotationNode)) {
 			return;
-		} else {
-			generateBuilderMethod = true;
 		}
-		
-		if (!checkName("buildMethodName", buildMethodName, annotationNode)) return;
-		if (!builderClassName.isEmpty()) {
-			if (!checkName("builderClassName", builderClassName, annotationNode)) return;
+
+		boolean generateBuildMethod = !buildMethodName.isEmpty();
+		if (generateBuildMethod && !checkName("++buildMethodName", buildMethodName, annotationNode)) {
+			return;
+		}
+
+		if (!builderClassName.isEmpty() && !checkName("builderClassName", builderClassName, annotationNode)) {
+			return;
 		}
 		
 		deleteAnnotationIfNeccessary(annotationNode, Builder.class, "lombok.experimental.Builder");
@@ -437,7 +437,7 @@ public class HandleBuilder extends JavacAnnotationHandler<Builder> {
 			makeSetterMethodsForBuilder(builderType, bfd, annotationNode, fluent, chain, accessForInners);
 		}
 		
-		{
+		if (generateBuildMethod) {
 			MemberExistsResult methodExists = methodExists(buildMethodName, builderType, -1);
 			if (methodExists == MemberExistsResult.EXISTS_BY_LOMBOK) methodExists = methodExists(buildMethodName, builderType, 0);
 			if (methodExists == MemberExistsResult.NOT_EXISTS) {
@@ -448,7 +448,7 @@ public class HandleBuilder extends JavacAnnotationHandler<Builder> {
 				}
 			}
 		}
-		
+
 		if (methodExists("toString", builderType, 0) == MemberExistsResult.NOT_EXISTS) {
 			java.util.List<Included<JavacNode, ToString.Include>> fieldNodes = new ArrayList<Included<JavacNode, ToString.Include>>();
 			for (BuilderFieldData bfd : builderFields) {