+++ test/transform/resource/before/BuilderWithNoBuildMethod.java
@@ -1,0 +1,5 @@
+import lombok.Builder;
+@Builder(buildMethodName = "")
+class BuilderWithNoBuildMethod {
+	private String a;
+}